corrected a bug when Object.equals method would be used instead of IntVariable.equals

+++ backends/src/net/sf/orcc/backends/c/quasistatic/scheduler/unrollers/ConfigurationAnalyzer.java
@@ -131,7 +131,7 @@ public class ConfigurationAnalyzer {
 				if (o2 instanceof IntVariable) {
 					v1.equals((IntVariable) o2);
 				} else if (o2 instanceof Integer) {
-					v1.equals((Integer) o2);
+					v1.equals(((Integer) o2).intValue());
 				} else {
 					break;
 				}
